requirements.txt
openai
numpy
torch
termcolor
chromadb
pydantic
requests
python-dotenv
